refactor(api): extract task approvals + lifecycle handlers
Continues the per-concern split of handler_tasks.go. Approval and
run-lifecycle handlers (8 routes) move to a sibling file so each
handler file holds one thing.

  handler_tasks.go:           2022 → 1618 LOC
  handler_tasks_lifecycle.go: new, 430 LOC

What moves to handler_tasks_lifecycle.go

  Approvals (3)
  - HandleTaskApprovals
  - HandleTaskApproval
  - HandleResolveTaskApproval

  Run-state transitions (5)
  - HandleCancelTaskRun
  - HandleRetryTaskRun
  - HandleResumeTaskRun
  - HandleContinueTaskRun
  - HandleRetryTaskRunFromTurn

All 8 handlers route into runner.{Resume,Reject,Retry,Continue,Cancel}*
methods; bundling them clusters the runner-coordination surface in
one file.

What stays in handler_tasks.go

  CRUD (HandleCreateTask, HandleTasks, HandleTask, HandleDeleteTask,
  HandleStartTask), runs/steps (HandleTaskRuns, HandleTaskRun,
  HandleTaskRunSteps, HandleTaskRunStep), stream + events
  (HandleTaskRunStream, HandleTaskRunEvents, HandleAppendTaskRunEvent,
  buildTaskRunStreamState, decodeTaskRunEventData), and the shared
  loaders/helpers (loadAuthorizedTask, loadAuthorizedTaskRun,
  buildTaskItem, etc.). Stream + events extraction is the next
  sibling PR.

The receiver type stays *Handler everywhere; no public API change;
mux registrations in server.go are unchanged. handler_tasks.go's
imports were unaffected; handler_tasks_lifecycle.go takes its own
encoding/json + fmt + log/slog + net/http + strings + time +
telemetry + pkg/types.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  gofmt -s -l .                 — clean
  go test ./... -race -count=1  — 1779 / 1779 pass
